refactor(typescript): isolate legacy factories
