+++ typescript/src/index.ts
@@ -1,13 +1,8 @@
-export * from "./blocks.js";
 export * from "./builder.js";
-export * from "./elements.js";
 export * from "./errors.js";
-export * from "./messages.js";
-export * from "./objects.js";
-export * from "./rich-text.js";
 export * from "./types.js";
 export * from "./validation.js";
-export * from "./views.js";
 export * from "./fluent/index.js";
+export * from "./legacy/index.js";
 
 export const specVersion = "1.1.0";
