+++ typescript/src/legacy/index.ts
@@ -1,0 +1,13 @@
+/**
+ * Compatibility exports for the pre-fluent TypeScript API.
+ *
+ * @deprecated Prefer the PascalCase fluent builders exported from the package
+ * root. These lowercase factories remain only for v2 compatibility and are
+ * scheduled for removal in v3.
+ */
+export * from "./blocks.js";
+export * from "./elements.js";
+export * from "./messages.js";
+export * from "./objects.js";
+export * from "./rich-text.js";
+export * from "./views.js";
